refactor: merged Content.svelte to +page.svelte for electricity dashboard
